backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
google-generativeai==0.7.2
google-api-python-client==2.137.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-multipart==0.0.9
python-dotenv==1.0.1
